test/DekstopApp/Cert_Tool/requirements.txt
certifi
requests
pillow
qrcode
cryptography
PyPDF2
pypdf
charset-normalizer
cffi
```